Separate domain layer with mobx store
