+++ src/domain/auth/authGateway.ts
@@ -1,0 +1,5 @@
+import { Token } from './entity';
+
+export interface AuthGateway {
+  login(req: { username: string; password: string }): Promise<Token>;
+}
