+++ src/domain/auth/authStorage.ts
@@ -1,4 +1,4 @@
-export interface KeyValueStorage {
+export interface AuthStorage {
   get(key: string): string | null;
   set(key: string, value: string): void;
   unset(key: string): void;
