+++ src/domain/auth/entity.ts
