--- src/domain/dto.ts
@@ -1,1 +1,0 @@
-export type Result<T, U> = [T?, U?];
