--- src/domain/index.ts
@@ -1,13 +1,0 @@
-import { AuthInteractor } from './auth/interactor';
-import { UserInteractor } from './user/interactor';
-import { TodoInteractor } from './todo/interactor';
-import { Result } from './dto';
-import { KeyValueStorage } from './persist/storage';
-
-export {
-  AuthInteractor,
-  UserInteractor,
-  TodoInteractor,
-  Result,
-  KeyValueStorage,
-};
