+++ src/domain/todo/entity.ts
@@ -1,0 +1,4 @@
+export interface Todo {}
+export interface TodoCategory {
+  readonly name: string;
+}
