+++ src/domain/todo/index.ts
@@ -1,0 +1,5 @@
+import { Todo, TodoCategory } from './entity';
+import { TodoGateway } from './todoGateway';
+import { TodoUsecase } from './usecase';
+
+export { Todo, TodoCategory, TodoGateway, TodoUsecase };
