+++ src/domain/todo/todoGateway.ts
@@ -1,0 +1,5 @@
+import { Todo } from './entity';
+
+export interface TodoGateway {
+  todos(): Promise<Todo[]>;
+}
