+++ src/domain/user/index.ts
@@ -1,0 +1,5 @@
+import { User } from './entity';
+import { UserGateway } from './userGateway';
+import { UserUsecase } from './usecase';
+
+export { User, UserGateway, UserUsecase };
