+++ src/domain/user/usecase.ts
@@ -1,0 +1,5 @@
+import { User } from './entity';
+
+export interface UserUsecase {
+  me(): Promise<User>;
+}
