+++ src/domain/user/userGateway.ts
@@ -1,0 +1,5 @@
+import { User } from '.';
+
+export interface UserGateway {
+  me(): Promise<User>;
+}
