+++ src/stores/userStore.ts
@@ -1,12 +1,8 @@
-import {
-  UserInteractor,
-  UserGateway,
-  UserResult,
-} from '@/domain/user/interactor';
+import { UserUsecase, UserGateway, UserResult } from '@/domain/user/usecase';
 import { Result } from '@/domain';
-import { User } from '@/domain/user/dto';
+import { User } from '@/domain/user/entity';
 
-export class UserStore implements UserInteractor {
+export class UserStore {
   constructor(readonly userGateway: UserGateway) {}
 
   async me(): Promise<Result<UserResult, Error>> {
